backend/requirements.txt
asgiref==3.5.2
Django==4.1.1
sqlparse==0.4.2
django-environ==0.9.0
psycopg2-binary==2.9.3
djangorestframework==3.14.0
django-cors-headers==3.13.0
drf-yasg==1.21.4
black==22.8.0
factory-boy==3.2.1
Faker
